Now it's possible to add vehicles from the page

+++ CarRental.Common/Classes/Motorcycle.cs
@@ -7,7 +7,7 @@ public class Motorcycle : Vehicle
 {
 
 
-    public Motorcycle(string regNo, string make, int odometer, double costKm, VehicleTypes vehicleType, VehicleStatuses vehicleStatus) :
+    public Motorcycle(string regNo, string make, double odometer, double costKm, VehicleTypes vehicleType, VehicleStatuses vehicleStatus) :
         base(regNo, make, odometer, costKm, vehicleType, vehicleStatus)
     { }
 }
