+++ CarRental.Common/Classes/Vehicle.cs
@@ -13,12 +13,12 @@ public abstract class Vehicle : IVehicle
     public int Id { get; set; }
     public string RegNo { get; init; }
     public string Make { get; init; }
-    public int Odometer { get; set; }
+    public double Odometer { get; set; }
     public double CostKm { get; init; }
     public VehicleTypes VehicleType { get; init; }
     public VehicleStatuses VehicleStatus { get; set; }
 
-    public Vehicle(string regNo, string make, int odometer, double costKm, VehicleTypes vehicleType, VehicleStatuses vehicleStatus)
+    public Vehicle(string regNo, string make, double odometer, double costKm, VehicleTypes vehicleType, VehicleStatuses vehicleStatus)
     {
         RegNo = regNo;
         Make = make;
